Biegunka.Source.Darcs

Description

Biegunka.Source.Darcs - functions to work with darcs repositories as sources

Source layer

darcs :: String -> FilePath -> Script Actions () -> Script Sources ()

Clone repository from the given url to specified path and/or pull. Also executes attached script

 darcs "https://example.com/repository" "darcs/repository" $ do
   registerAt "some/not/so/long/path"
   link "important.file" ".config"

get repository from https://example.com/repository to ${HOME}/darcs/repository

link ${HOME}/darcs/repository to ${HOME}/some/not/so/long/path

link ${HOME}/darcs/repository/important.file to ${HOME}/.config

darcs_ :: String -> FilePath -> Script Sources ()

Clone repository from the given url to specified path

 darcs_ "https://example.com/repository" "darcs/repository"

get repository from https://example.com/repository to ${HOME}/darcs/repository